requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
gunicorn==22.0.0
itsdangerous==2.2.0
Werkzeug==3.0.3
Jinja2==3.1.4
# （将来 Postgres を使うなら）
# psycopg2-binary==2.9.9
# マイグレーションを使うなら
# Flask-Migrate==4.0.5
Flask-Mail==0.10.0
